_scripts/requirements.txt
arrow==1.3.0
attrs==25.3.0
ics==0.7.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
six==1.17.0
TatSu==5.13.1
types-python-dateutil==2.9.0.20241206
